+++ spec/orm/cases/default_values_spec.rb
@@ -16,6 +16,7 @@ describe "default values" do
     its(:dictionary) { should eq Hash.new }
     its(:points) { should eq 1 }
     its(:status) { should eq false }
+    its(:positive_as_default) { should eq true }
     its(:phone_numbers) { should eq [] }
   end
 
@@ -24,6 +25,7 @@ describe "default values" do
       subject.dictionary = nil
       subject.points = nil
       subject.status = nil
+      subject.positive_as_default = false
       subject.phone_numbers = nil
       subject.save!
       subject.reload
@@ -32,6 +34,7 @@ describe "default values" do
     its(:dictionary) { should be_nil }
     its(:points) { should be_nil }
     its(:status) { should be_nil }
+    its(:positive_as_default) { should be_false }
     its(:phone_numbers) { should eq [] }
   end
 end